Remove interface definitions from Android code
