+++ client/internal/routemanager/manager_android.go
@@ -8,12 +8,6 @@ import (
 	"github.com/netbirdio/netbird/route"
 )
 
-// Manager is a route manager interface
-type Manager interface {
-	UpdateRoutes(updateSerial uint64, newRoutes []*route.Route) error
-	Stop()
-}
-
 // DefaultManager dummy router manager for Android
 type DefaultManager struct {
 	ctx          context.Context
